clean one-time usage code which fixes DB. split out resource quota computation code
