+++ src/utils/MySQLDataHandler.py
@@ -8,6 +8,7 @@ import timeit
 
 import mysql.connector
 from prometheus_client import Histogram
+from vc_quota import vc_value_str
 
 from config import config, global_vars
 
@@ -42,169 +43,6 @@ def base64encode(str_val):
 
 def base64decode(str_val):
     return base64.b64decode(str_val.encode("utf-8")).decode("utf-8")
-
-
-def get_vc_resource_ratio(config):
-    if "vc_gpu_quota" in config:
-        ratio = get_ratio_from_config(config)
-    else:
-        ratio = query_vc_4_ratio(config)
-    print("ratio is:\n")
-    print(ratio)
-    return ratio
-
-
-def get_ratio_from_config(config):
-    total = 0
-    ratio = {}
-    for vc_val in config["vc_gpu_quota"].values():
-        for num in vc_val.values():
-            if isinstance(3, int):
-                total += num
-    for vc, vc_val in config["vc_gpu_quota"].items():
-        tmp_sum = 0
-        for num in vc_val.values():
-            if isinstance(3, int):
-                tmp_sum += num
-        ratio[vc] = tmp_sum / total
-    return ratio
-
-
-def connect2DB(config):
-    database = "DLWSCluster-%s" % config["clusterId"]
-    server = config["mysql"]["hostname"]
-    username = config["mysql"]["username"]
-    password = config["mysql"]["password"]
-    conn = mysql.connector.connect(user=username, password=password, host=server, database=database)
-    return conn
-
-
-def query_vc_4_ratio(config):
-    conn = connect2DB(config)
-    mycursor = conn.cursor()
-    mycursor.execute("SELECT * FROM vc;")
-    myresult = mycursor.fetchall()
-    total = 0
-    for entry in myresult:
-        total += sum(eval(entry[4]).values())
-    ratio = {}
-    for entry in myresult:
-        ratio[str(entry[2])] = sum(eval(entry[4]).values()) / total
-    return ratio
-
-
-def update_db(config):
-    ratio_dict = get_vc_resource_ratio(config)
-    _, _, res_quota, res_meta = vc_value_str(config, ratio_dict)
-    conn = connect2DB(config)
-    sqls = []
-    if not column_exists(conn, 'vc', 'resourceQuota'):
-        sqls = ["ALTER TABLE vc ADD COLUMN `resourceQuota` TEXT NOT NULL AFTER `metadata`;", 
-            "ALTER TABLE vc ADD COLUMN `resourceMetadata` TEXT NOT NULL AFTER `resourceQuota`;"]
-    for vc, vc_res in res_quota.items():
-        sqls.append("UPDATE vc SET resourceQuota = '{}' where vcName = '{}'".format(vc_res, vc).replace('"', '\\\"'))
-        sqls.append("UPDATE vc SET resourceMetadata = '{}' where vcName = '{}'".format(res_meta, vc).replace('"', '\\\"'))
-    print(sqls)
-    execute_mysql(conn, sqls)
-    return sqls
-
-
-def column_exists(conn, tablename, columnname):
-    sql = "SHOW COLUMNS FROM `{}` LIKE '{}';".format(tablename, columnname)
-    cursor = conn.cursor()
-    cursor.execute(sql)
-    result = cursor.fetchone()
-    print(result)
-    return not result is None
-
-
-def table_exists(conn, tablename):
-    sql = "SHOW TABLES LIKE '{}'".format(tablename)
-    cursor = conn.cursor()
-    cursor.execute(sql)
-    result = cursor.fetchone()
-    print(result)
-    return not result is None
-
-def execute_mysql(conn, sqls):
-    if isinstance(sqls, str):
-        sqls = [sqls]
-    for sql in sqls:
-        cursor = conn.cursor()
-        cursor.execute(sql)
-        conn.commit()
-        cursor.close()
-
-
-def vc_value_str(config, ratio_dict):
-    if "worker_sku_cnt" not in config or "sku_mapping" not in config:
-        print("Warning: no default value would be added to VC table. Need to manually specify")
-        return "", "", [], ""
-
-    worker_sku_cnt, sku_mapping = config["worker_sku_cnt"], config["sku_mapping"]
-    quota_dict = {}
-    old_meta = {}
-    resource_quota = {"cpu":{}, "memory":{}, "gpu":{}, "gpu_memory":{}}
-    for sku, cnt in worker_sku_cnt.items():
-        gpu_type = sku_mapping.get(sku, {}).get("gpu-type", "None")
-        num_gpu_per_node = sku_mapping.get(sku, {}).get("gpu", 0)
-        quota_dict[gpu_type] = quota_dict.get(gpu_type, 0) + cnt * num_gpu_per_node
-        old_meta[gpu_type] = {"num_gpu_per_node": num_gpu_per_node}
-        sku_name_in_map = sku if sku in sku_mapping else ""
-        meta_tmp = sku_mapping.get(sku_name_in_map, {})
-        for r_type in resource_quota.keys():
-            resource_quota[r_type][sku_name_in_map] = resource_quota[r_type].get(sku_name_in_map, 0) + meta_tmp.get(r_type, 0) * cnt
-
-    for r_type in ["cpu", "memory"]:
-        for sku, val in resource_quota[r_type].items():
-            resource_quota[r_type][sku] *= config.get("vc_rsc_discount", 0.9)
-
-    # default value of quota and metadata are based on the assumption that there's only one default VC, this is not reasonable, and 
-    # these 2 fields would also finally get removed
-    quota = json.dumps(quota_dict, separators=(",", ":"))
-    metadata = json.dumps(old_meta, separators=(",", ":"))
-
-    res_quota = {}
-    for vc, ratio in ratio_dict.items():
-        res_quota_dict = {}
-        for res, res_q in resource_quota.items():
-            tmp_res_quota = {}
-            for sku, cnt in res_q.items():
-                cnt_p = cnt * ratio
-                if "memory" in res:
-                    cnt_p = '{}Gi'.format(cnt)
-                tmp_res_quota[sku] = cnt_p
-            res_quota_dict[res] = tmp_res_quota
-        res_quota[vc] = json.dumps(res_quota_dict, separators=(",", ":"))
-
-    res_meta_dict = {}
-    for r_type in ["cpu", "memory", "gpu", "gpu_memory"]:
-        tmp_res_meta = {}
-        for sku in worker_sku_cnt:
-            sku_name_in_map = sku if sku in sku_mapping else ""
-            pernode_cnt = sku_mapping.get(sku_name_in_map, {}).get(r_type, 0)
-            if "memory" in r_type:
-                pernode_cnt = '{}Gi'.format(pernode_cnt)
-            tmp_res_meta[sku_name_in_map] = {"per_node": pernode_cnt}
-            if r_type == "gpu":
-                tmp_res_meta[sku_name_in_map]["gpu_type"] = sku_mapping.get(sku_name_in_map, {}).get("gpu-type", "None")
-        res_meta_dict[r_type] = tmp_res_meta
-    res_meta = json.dumps(res_meta_dict, separators=(",", ":"))
-    return quota, metadata, res_quota, res_meta
-
-def insert_vc_sqls(config):
-    ratio_dict = {config['defalt_virtual_cluster_name']: 1.0}
-    if "vc_resource_ratio" in config:
-        ratio_dict = config["vc_resource_ratio"]
-    quota, metadata, res_quota, res_meta = vc_value_str(config, ratio_dict)
-    if quota == "":
-        return ""
-    sqls = []
-    for i, (vc, vc_res_quota) in enumerate(res_quota.items()):
-        sql = "INSERT INTO `vc` (`id`, `vcName`, `parent`, `quota`, `metadata`, `resourceQuota`, `resourceMetadata`) VALUES ({}, '{}', NULL, '{}', '{}', '{}', '{}')".format(
-            i+1, vc, quota, metadata, vc_res_quota, res_meta)
-        sqls.append(sql)
-    return sqls
 
 
 class DataHandler(object):
@@ -353,8 +191,6 @@ class DataHandler(object):
             # when the VC has vm of same GPU type but different VMsizes, e.g., when VC has Standard_NC6s_v3 and Standard_NC12s_v3 both?
             # impossible since there's no way to do it with current config mechanism
 
-            vc_table_exist = table_exists(self.conn, self.vctablename)
-
             sql = """
                 CREATE TABLE IF NOT EXISTS  `%s`
                 (
@@ -376,9 +212,7 @@ class DataHandler(object):
             self.conn.commit()
             cursor.close()
 
-            if not vc_table_exist:
-                defalult_vc_rows = insert_vc_sqls(config)
-                execute_mysql(self.conn, defalult_vc_rows)
+            self.init_vc_sqls(config)
 
 
 
@@ -440,6 +274,23 @@ class DataHandler(object):
             self.conn.commit()
             cursor.close()
 
+    @record
+    def column_exists(self, tablename, columnname):
+        sql = "SHOW COLUMNS FROM `{}` LIKE '{}';".format(tablename, columnname)
+        cursor = self.conn.cursor()
+        cursor.execute(sql)
+        result = cursor.fetchone()
+        print(result)
+        return result is not None
+
+    @record
+    def table_exists(self, tablename):
+        sql = "SHOW TABLES LIKE '{}'".format(tablename)
+        cursor = self.conn.cursor()
+        cursor.execute(sql)
+        result = cursor.fetchone()
+        print(result)
+        return result is not None
 
     @record
     def AddStorage(self, vcName, url, storageType, metadata, defaultMountPath):
@@ -503,13 +354,25 @@ class DataHandler(object):
             logger.error('Exception: %s', str(e))
             return False
 
-
     @record
-    def AddVC(self, vcName, quota, metadata):
+    def init_vc_sqls(self, config):
+        ratio_dict = {config['defalt_virtual_cluster_name']: 1.0}
+        if "vc_resource_ratio" in config:
+            ratio_dict = config["vc_resource_ratio"]
+        quota, metadata, res_quota, res_meta = vc_value_str(config, ratio_dict)
+        if quota == "":
+            return
+        sqls = []
+        for i, (vc, vc_res_quota) in enumerate(res_quota.items()):
+            print(vc, quota, metadata, vc_res_quota, res_meta)
+            self.AddVC(vc, quota, metadata, vc_res_quota, res_meta)
+
+    def AddVC(self, vcName, quota, metadata, res_quota, res_meta):
         try:
-            sql = "INSERT INTO `"+self.vctablename+"` (vcName, quota, metadata) VALUES (%s,%s,%s)"
+            sql = "INSERT INTO `{}` (vcName, quota, metadata, resourceQuota, resourceMetadata) VALUES ('{}', '{}', '{}', '{}', '{}')".format(self.vctablename, vcName, quota, metadata, res_quota, res_meta)
+            print(sql)
             cursor = self.conn.cursor()
-            cursor.execute(sql, (vcName, quota, metadata))
+            cursor.execute(sql)
             self.conn.commit()
             cursor.close()
             return True
